ignorefs: moved ignoring cache directories to ignorefs layer
